Logger cleanup
