+++ src/main/java/com/gtnewhorizons/angelica/AngelicaMod.java
@@ -6,7 +6,6 @@ import cpw.mods.fml.common.Mod;
         modid = "angelica",
         name = "Angelica",
         version = Tags.VERSION,
-        acceptedMinecraftVersions = "[1.7.10,)",
+        acceptedMinecraftVersions = "[1.7.10]",
         acceptableRemoteVersions = "*")
-public class AngelicaMod {
-}
+public class AngelicaMod {}
