+++ src/main/java/com/gtnewhorizons/angelica/loading/AngelicaTweaker.java
@@ -12,6 +12,7 @@ import org.spongepowered.asm.launch.GlobalProperties;
 import org.spongepowered.asm.service.mojang.MixinServiceLaunchWrapper;
 
 import com.gtnewhorizon.gtnhmixins.IEarlyMixinLoader;
+import com.gtnewhorizons.angelica.ALog;
 import com.gtnewhorizons.angelica.mixins.Mixins;
 
 import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
@@ -19,8 +20,6 @@ import cpw.mods.fml.relauncher.IFMLLoadingPlugin;
 @IFMLLoadingPlugin.MCVersion("1.7.10")
 @IFMLLoadingPlugin.SortingIndex(1100)
 public class AngelicaTweaker implements IFMLLoadingPlugin, IEarlyMixinLoader {
-
-    public static final Logger log = LogManager.getLogger("AngelicaTweaker");
 
     @Override
     public String[] getASMTransformerClass() {
@@ -69,7 +68,7 @@ public class AngelicaTweaker implements IFMLLoadingPlugin, IEarlyMixinLoader {
                 }
             }
         }
-        log.info("Not loading the following EARLY mixins: {}", notLoading.toString());
+        ALog.info("Not loading the following EARLY mixins: {}", notLoading.toString());
         return mixins;
     }
 }
