+++ src/main/java/com/gtnewhorizons/angelica/loading/MixinCompatHackTweaker.java
@@ -9,8 +9,8 @@ import net.minecraft.launchwrapper.ITweaker;
 import net.minecraft.launchwrapper.Launch;
 import net.minecraft.launchwrapper.LaunchClassLoader;
 
+import com.gtnewhorizons.angelica.ALog;
 import com.gtnewhorizons.angelica.transform.AClassTransformer;
-import com.gtnewhorizons.angelica.transform.ALog;
 
 import cpw.mods.fml.common.asm.transformers.TerminalTransformer;
