+++ src/main/java/com/gtnewhorizons/angelica/transform/ACTItemRenderer.java
@@ -11,6 +11,8 @@ import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class ACTItemRenderer implements IClassTransformer {
 
     @Override
