+++ src/main/java/com/gtnewhorizons/angelica/transform/ACTRendererLivingEntity.java
@@ -13,6 +13,8 @@ import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class ACTRendererLivingEntity implements IClassTransformer {
 
     @Override
