+++ src/main/java/com/gtnewhorizons/angelica/transform/ACTTextureDownload.java
@@ -12,6 +12,8 @@ import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class ACTTextureDownload implements IClassTransformer {
 
     @Override
