+++ src/main/java/com/gtnewhorizons/angelica/transform/ACTTextureDynamic.java
@@ -10,6 +10,8 @@ import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class ACTTextureDynamic implements IClassTransformer {
 
     @Override
