+++ src/main/java/com/gtnewhorizons/angelica/transform/ACTTextureLayered.java
@@ -11,6 +11,8 @@ import org.objectweb.asm.Label;
 import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Opcodes;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class ACTTextureLayered implements IClassTransformer {
 
     @Override
