+++ src/main/java/com/gtnewhorizons/angelica/transform/AClassTransformer.java
@@ -9,6 +9,8 @@ import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.tree.ClassNode;
 
+import com.gtnewhorizons.angelica.ALog;
+
 public class AClassTransformer implements IClassTransformer {
 
     Names names;
